Segunda iteracion, ahora se utiliza un nuevo componenete con su propia interfaz y logica.
